Move theme style files to subfolder & update refs
