+++ src/components/App.css
@@ -1,4 +1,4 @@
-@value themeBackdrop from "theme.module.css";
+@value themeBackdrop from "theme/theme.module.css";
 
 body {
 	background: themeBackdrop;
