+++ src/components/LegacyAnalyse/SegmentLinkItem.module.css
@@ -1,4 +1,4 @@
-@value gutter from "theme.module.css";
+@value gutter from "theme/theme.module.css";
 
 .link {
 	margin: 0 calc(-1 * gutter);
