+++ src/parser/core/modules/ResourceGraphs/ResourceGraphs.module.css
@@ -1,4 +1,4 @@
-@value themeBackground from 'theme.module.css';
+@value themeBackground from 'theme/theme.module.css';
 
 @value markerPointerSize: 5px;
