+++ src/parser/core/modules/Suggestions/Suggestions.module.css
@@ -1,4 +1,4 @@
-@value gutter from "theme.module.css";
+@value gutter from "theme/theme.module.css";
 
 .checkbox {
 	position: absolute !important;
