+++ src/parser/core/modules/Timeline/Timeline.module.css
@@ -1,4 +1,4 @@
-@value gutter from "theme.module.css";
+@value gutter from "theme/theme.module.css";
 
 .helpText {
 	font-style: italic;
