+++ src/parser/jobs/ast/modules/Sect/SectStatistic.module.css
@@ -1,4 +1,4 @@
-@value gutter from "theme.module.css";
+@value gutter from "theme/theme.module.css";
 
 .simpleStatistic {
 	position: relative;
