+++ src/parser/jobs/rdm/statistics/DualStatistic.module.css
@@ -1,4 +1,4 @@
-@value gutter from "theme.module.css";
+@value gutter from "theme/theme.module.css";
 
 .dualStatistic {
 	position: relative;
